requirements.txt
openai>=1.12.0
pillow>=10.0.0
mss>=9.0.1
python-dotenv>=1.0.0
rich>=13.7.0
pynput>=1.7.6
markdown>=3.5.1

